add experimental support for windows
this has come up once or twice as there aren't many good libraries
for writing cross-platform terminal applications. to my surprise,
recent windows versions support xterm-like caps if you put the
console in a special mode. initializing the console, reading and
writing to it, and handling resize events are different but
everything else is the same.

`wcwidth` isn't available, so everything is assumed to occupy 1
column.

i haven't tested anything besides the keyboard demo, which seems to
work just fine.