fix bug with adding combining chars to a full-width char
the logic was incorrectly sending the combinin char to x-1, but that
assumed a wcwidth==1 char. save last printed x coord and use that
instead.

xterm's ansi output for this is a little strange. unlike other
`expected.ansi` files, cat-ing this one to stdout doesn't seem to
render the correct output, and if you peek at its contents you'll see
some "\xef\xbf\xbf" (U+FFFF) which is an invalid codepoint. in any
case, it still picks up the regression so i'm adding it.