set `global.ttyfd` if either `rfd` or `wfd` `isatty`
normally, termbox initializes itself by opening `/dev/tty`, but one
can also initialize it with arbitrary read and write fds via
`tb_init_rwfd`. while this is mainly useful for testing (e.g.,
`tests/test_mouse`), it may have real world applications as well.

for example, one might want to read input from a pipe and render to a
tty. presently that isn't possible as `tb_init_rwfd` only sets
`global.ttyfd` if `rfd == wfd && isatty(rfd)`. presence of
`global.ttyfd` gates execution of various `ioctl`s, like putting the
terminal in raw mode, which only makes sense to call if we know the
fd is a tty.

in our hypothetical case, the read and write fds are different, and
only the write fd is a tty, which we still want to `ioctl`. this
patch makes it possible to do this.

the only odd case is if `rfd` and `wfd` are different fds and both
ttys. in this case, `rfd` takes precedence (arbitrarily). termbox
does not support initializing multiple ttys. (please do get in
touch if you're looking to do that because it sounds cool.)