add experimental windows support via `windows.sh`
windows support has been requested several times throughout the
years, however i've always been reluctant to pollute the library with
ifdefs and/or abstraction layers for this sake. we avoid this with a
shell script (`windows.sh`) that does surgery on `termbox2.h` to
generate a windows-compatible version. some headers are swapped, some
functions are altered or replaced, but most of the library stays the
same. a couple marker comments (`__posix_(start|end)`) are added to
`termbox2.h` which `windows.sh` uses to anchor various sed/awk
commands.

theoretically the test suite could run on windows. it would take some
research to figure out the windows equivalents of things in
`tests/run.sh` like `xvkbd` etc. github docs state that
`windows-latest` has php pre-installed, and php does have windows ffi
support. the memfd-based tests would have to be skipped (or proper
rwfd support would need to be added to the windows lib).

instead of running the test suite, we ensure we can at least compile
the keyboard demo on windows. in my opinion this suffices for
experimental support.