fix grapheme cluster width
as pointed out in #125, the width of a grapheme cluster is not the
sum of the wcwidth of its codepoints. there's no standard way of
determining the column width of a grapheme cluster, but the
implementation in this commit is a decent real world approximation.
basically, it returns the max wcwidth in the cluster, and tries to
account for vs15, vs16, zero-width joiners, and regional indicators.